requirements.txt
asgiref==3.2.8
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-credit-cards==0.4.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
